+++ packages/uiengine-theme-default/src/vue/components/AppNavigationTree.vue
@@ -11,6 +11,7 @@
       :item="navigation[itemId]"
       :level="level"
     />
+    <slot />
   </ul>
 </template>