+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Can be used to cross-reference and search items
 - Ability for toggling the navigation on desktop breakpoints
 - Links in footer to navigate to the previous and next page
+- Settings for locale and highlight.js theme
 - Integration for Google Analytics
   - Provide your `analyticsId` (`UA-XXX-X`) in the [project config](./docs/config.md)